+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Improvements:
 * Update the bundled MLKit model for Android to version `17.3.0`.
 * Added documentation in places where it was missing.
 * Added `color` and `style` properties to the `BarcodePainter` widget.
+* Enabled Swift Package Manager for the example app.
 
 ## 7.0.0-beta.3