+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## NEXT
+## 7.0.0-beta.4
 
 **BREAKING CHANGES:**